requirements.txt
tensorflow==2.11.1
tqdm
scikit-learn
biopython==1.82
numpy